BREAKING: Split core and plattform implementation With this commit MyWarp has a platform independent core that does not use any platform specific code, and a platform specific implementation that adapts this core to the logic used by the platform running it. As of this commit, only Bukkit is supported and the implementation is not yet splitted into another Maven module.
There are also some functional differences that should be noted:
- The names of the SQL tables, columns and constraints have changed.
- The world is now identified and stored by UUID.
- Additional logic to lookup player-names have been added on a per-plattform-level. Fixes #15.
- The creation-date of a Warp is now stored in the database and used by
the popularity algorithm, this fixes #10.
- The internationalization has been rewritten and uses standard
Properties resource bundles. For now they are not accessible to the
end-user.
- Formatting has been completely removed from the internationalized
resources and is hard-coded to make resources platform independent.
- The custom command management has been replaced by Intake, this fixes #11.
- The ‚-p‘ flag for list and search is gone.
- The list command has been altered, its ‚-c‘ and ‚-w‘ flags check if
the given query is contained in the name of the creator or world.
- A simple internal event system for warps has been added that uses
Guava’s EventBus to notify subscribers of changes. It is currently used
to notify that data layer and Dynmap, if enabled.
- Each class that needs to log something now uses its own logger. These
loggers currently do not append the plugin’s name as they probably
should.

There also several aspects that still need work. Most of them are
documented right in the source. It is fully expected that the overall
design will change again as there are still some problems left to solve.

This commit has only been briefly tested. Standard functions work, but
it is expected that there are bugs or unintended behavior. By no means
this is a stable product!