Add Tab complete for Warps, LocalPlayers and Profiles
In terms of functionality, this commit is pretty much experimental: Warps, LocalPlayers and Profiles all use the matching logic that was previously used to parse warp names (see documentation of Matcher for details). Suggestions on Tab are ordered alphabetically (LocalPlayers, Profiles) or by popularity (Warps). Due to the behavior of Intake, Tab complete currently does not work for warps when the '/mywarp <warp>' command is used as it parses the sub-commands instead.

This commit requires that latest version of my Intake fork. Use '--refresh-dependencies' if the build fails.

Closes #107, #83.