Add automatic download of translations from crowdin upon build
To download the translations, crowdin's API key is required. It can be set when launching the build by using the 'crowdinKey' property like so: 'gradlew clean build -PcrowdinKey=PRIVATE-API-KEY'. If this property is not set, the build continues with the bundled translations only.

The German translation is no longer bundled directly by MyWarp as it is hosted on crowdin and bundled upon build.

Fixes #109.