Fix a possible crash for malformatted value bundles, use correct database path when falling back to defaults
If a value bundle configuration is not a valid ConfigurationSection, the values are ignored and a warning is logged. If a value bundle's values are missing, defaults are used.

If MyWarp fails to create the configuration file and falls back to build-in defaults, the default path to the H2 database is now used correctly.

Fixes #102.