Move support for physical warp signs into the core
As before, implementations are responsible for listening to the relevant events and filtering these events for relevant interactions. These interactions are than given to the core to handle them, that is to search for warp signs associated with them.
While this implementations requires API users to manage relevant interactions - that is interactions with relevant blocks - by themselves, this should be faster, especially on platforms like Sponge that allow some native event filtering based on involved blocks.

Fixes #84.