Handle warps with non-existing worlds more gracefully
Before this commit, trying to get a no longer existing LocalWorld from a Warp resulted in a RuntimeException that was or was not handled, depending on the interaction model. As of this commit, the callers need to convert the stored world identifier to a LocalWorld by themselves, giving them control over the response if the world does not exist:
- Teleports to warps with non-exting worlds are apported and the user is informed accordingly.
- The 'info' and 'list' commands both display the stored UUID if the world does not exist.
- The 'give', 'public' and 'private' commands fail if the warp's world is unavailble and limits must checked (limits can only be checked for existing worlds).
- If world access control is enabled, Warps without an existing world cannot be viewed or used (if the world does not exist, there is no way to check if the user would be able to use the warp).
- If Dynmap is used, warps with non-exting worlds are ignored.

Additionally the handling of world access control and permission based authentication has been changed. Before this commit, world access control could overrule permission overrides; as of this commit, permission overrides overrule world access control.

Fixes #93.